+++ kommand-api/src/main/kotlin/io/github/monun/kommand/node/KommandNode.kt
@@ -22,12 +22,16 @@ import io.github.monun.kommand.*
 
 // 커맨드 노드
 @KommandDSL
-interface KommandNode: KommandArgumentSupport {
+interface KommandNode : KommandArgumentSupport {
     fun requires(requires: KommandSource.() -> Boolean)
 
     fun executes(executes: KommandSource.(KommandContext) -> Unit)
 
     fun then(name: String, init: LiteralNode.() -> Unit)
 
-    fun then(argument: Pair<String, KommandArgument<*>>, vararg arguments: Pair<String, KommandArgument<*>>, init: ArgumentNode.() -> Unit)
+    fun then(
+        argument: Pair<String, KommandArgument<*>>,
+        vararg arguments: Pair<String, KommandArgument<*>>,
+        init: ArgumentNode.() -> Unit
+    )
 }