+++ kommand-api/src/main/kotlin/io/github/monun/kommand/wrapper/WrapperSupport.kt
@@ -21,7 +21,7 @@ package io.github.monun.kommand.wrapper
 import io.github.monun.kommand.loader.LibraryLoader
 
 interface WrapperSupport {
-    companion object: WrapperSupport by LibraryLoader.loadNMS(WrapperSupport::class.java)
+    companion object : WrapperSupport by LibraryLoader.loadNMS(WrapperSupport::class.java)
 
     fun entityAnchorFeet(): EntityAnchor