+++ kommand-core/v1.17/src/main/kotlin/io/github/monun/kommand/v1_17/NMSKommand.kt
@@ -23,6 +23,7 @@ import com.mojang.brigadier.arguments.ArgumentType
 import com.mojang.brigadier.builder.ArgumentBuilder
 import com.mojang.brigadier.builder.LiteralArgumentBuilder
 import com.mojang.brigadier.builder.RequiredArgumentBuilder
+import com.mojang.brigadier.exceptions.CommandSyntaxException
 import com.mojang.brigadier.tree.CommandNode
 import com.mojang.brigadier.tree.LiteralCommandNode
 import com.mojang.brigadier.tree.RootCommandNode
@@ -77,12 +78,20 @@ private fun AbstractKommandNode.convert(): ArgumentBuilder<CommandSourceStack, *
     }.apply {
         requires?.let { requires ->
             requires { source ->
-                wrapSource(source).requires()
+                wrapSource(source).runCatching {
+                    requires()
+                }.onFailure {
+                    if (it !is CommandSyntaxException) it.printStackTrace()
+                }.getOrThrow()
             }
         }
         executes?.let { executes ->
             executes { context ->
-                wrapSource(context.source).executes(wrapContext(context))
+                wrapSource(context.source).runCatching {
+                    executes(this@convert.wrapContext(context))
+                }.onFailure {
+                    if (it !is CommandSyntaxException) it.printStackTrace()
+                }.getOrThrow()
                 1
             }
         }