+++ kommand-core/v1.17/src/main/kotlin/io/github/monun/kommand/v1_17/NMSKommandArgument.kt
@@ -467,7 +467,8 @@ class NMSKommandArgumentSupport : KommandArgumentSupport {
         function: KommandSource.(context: KommandContext, input: String) -> T?
     ): KommandArgument<T> {
         return type.createType() provideDynamic { context, name ->
-            context.source.function(context, name) ?: throw unknownArgument.create()
+            context.source.function(context, StringArgumentType.getString(context.handle, name))
+                ?: throw unknownArgument.create()
         }
     }
 }