+++ kommand-core/v1.17/src/main/kotlin/io/github/monun/kommand/v1_17/NMSKommandSource.kt
@@ -47,7 +47,7 @@ class NMSKommandSource private constructor(
             }
     }
 
-    internal val handle by weak(handle)
+    private val handle by weak(handle)
 
     override val displayName: Component
         get() = PaperBrigadier.componentFromMessage(handle.displayName)